pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "jwt>=1.3.1",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]
